Refactor dataset orchestration and tighten generation docs
